apps/api/pyproject.toml
[project]
name = "auto-cae-api"
version = "0.0.1"
description = "Auto-CAE backend: STEP loader, meshing, CalculiX runner"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "websockets>=13",
  "pydantic>=2.8",
  "numpy>=1.26",
  "python-multipart>=0.0.9",
  "cadquery-ocp>=7.7.2",
  "gmsh>=4.12",
]

[tool.ruff]
line-length = 100
target-version = "py311"
